feat: improve history page and rest page
