+++ src/components/History/History.tsx
@@ -24,53 +24,56 @@ const History = () => {
   }, []);
 
   return (
-    <div className={history__container}>
-      {history.length > 0 ? (
-        history
-          .slice()
-          .reverse()
-          .map((item, index) => {
-            const fullUrl = item.url;
+    <>
+      <h1 className='maintext maintext_green'>{t('history')}</h1>
+      <div className={history__container}>
+        {history.length > 0 ? (
+          history
+            .slice()
+            .reverse()
+            .map((item, index) => {
+              const fullUrl = item.url;
 
-            const headersParams = new URLSearchParams();
-            item.headers.forEach((header) => {
-              if (header.key && header.value) {
-                headersParams.append(header.key, header.value);
-              }
-            });
-            const methodColor =
-              HTTP_METHODS.find((m) => m.value === item.method)?.color || '';
+              const headersParams = new URLSearchParams();
+              item.headers.forEach((header) => {
+                if (header.key && header.value) {
+                  headersParams.append(header.key, header.value);
+                }
+              });
+              const methodColor =
+                HTTP_METHODS.find((m) => m.value === item.method)?.color || '';
 
-            return (
-              <Link
-                key={index}
-                href={fullUrl}
-                className=''
-              >
-                <div className={history__item}>
-                  <div>
-                    {t('date')} {item.date}
+              return (
+                <Link
+                  key={index}
+                  href={fullUrl}
+                  className=''
+                >
+                  <div className={history__item}>
+                    <div>
+                      {t('date')} {item.date}
+                    </div>
+                    <div>
+                      {t('method')}{' '}
+                      <span
+                        className={history__item_method}
+                        style={{ color: methodColor }}
+                      >
+                        {item.method}
+                      </span>
+                    </div>
+                    <div>
+                      {t('url')} {atob(item.requestURL)}
+                    </div>
                   </div>
-                  <div>
-                    {t('method')}{' '}
-                    <span
-                      className={history__item_method}
-                      style={{ color: methodColor }}
-                    >
-                      {item.method}
-                    </span>
-                  </div>
-                  <div>
-                    {t('url')} {atob(item.requestURL)}
-                  </div>
-                </div>
-              </Link>
-            );
-          })
-      ) : (
-        <h3>{t('noItems')}</h3>
-      )}
-    </div>
+                </Link>
+              );
+            })
+        ) : (
+          <h3>{t('noItems')}</h3>
+        )}
+      </div>
+    </>
   );
 };
